+++ test/packages/Customizable-Bootstrap-Stylus/module-definitions.js
@@ -1,0 +1,75 @@
+moduleDefinitions = {
+  "utilities":{stylus:["bootstrap/bootstrap-stylus/bootstrap/utilities.styl"]},
+  "responsive-utilities":{stylus:["bootstrap/bootstrap-stylus/bootstrap/responsive-utilities.styl"]},
+
+  "variables":{stylus:["bootstrap/bootstrap-stylus/bootstrap/variables.styl"]},
+  "mixins":{stylus:["bootstrap/bootstrap-stylus/bootstrap/mixins.styl"]},
+
+  "normalize":{stylus:["bootstrap/bootstrap-stylus/bootstrap/normalize.styl"]},
+  "print":{stylus:["bootstrap/bootstrap-stylus/bootstrap/print.styl"]},
+
+  "scaffolding":{stylus:["bootstrap/bootstrap-stylus/bootstrap/scaffolding.styl"]},
+  "type":{stylus:["bootstrap/bootstrap-stylus/bootstrap/type.styl"]},
+  "code":{stylus:["bootstrap/bootstrap-stylus/bootstrap/code.styl"]},
+  "grid":{stylus:["bootstrap/bootstrap-stylus/bootstrap/grid.styl"]},
+  "tables":{stylus:["bootstrap/bootstrap-stylus/bootstrap/tables.styl"]},
+  "forms":{stylus:["bootstrap/bootstrap-stylus/bootstrap/forms.styl"]},
+  "buttons":{stylus:["bootstrap/bootstrap-stylus/bootstrap/buttons.styl"]},
+
+  "glyphicons":{stylus:["bootstrap/bootstrap-stylus/bootstrap/glyphicons.styl"]},
+  "button-groups":{stylus:["bootstrap/bootstrap-stylus/bootstrap/button-groups.styl"]},
+  "input-groups":{stylus:["bootstrap/bootstrap-stylus/bootstrap/input-groups.styl"]},
+  "navs":{stylus:["bootstrap/bootstrap-stylus/bootstrap/navs.styl"]},
+  "navbar":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/utilities.styl","bootstrap/bootstrap-stylus/bootstrap/forms.styl","bootstrap/bootstrap-stylus/bootstrap/navs.styl","bootstrap/bootstrap-stylus/bootstrap/navbar.styl"],
+    js:["bootstrap/bootstrap-stylus/js/transition.js","bootstrap/bootstrap-stylus/js/collapse.js"]
+  },
+  "breadcrumbs":{stylus:["bootstrap/bootstrap-stylus/bootstrap/breadcrumbs.styl"]},
+  "pagination":{stylus:["bootstrap/bootstrap-stylus/bootstrap/pagination.styl"]},
+  "pager":{stylus:["bootstrap/bootstrap-stylus/bootstrap/pager.styl"]},
+  "labels":{stylus:["bootstrap/bootstrap-stylus/bootstrap/labels.styl"]},
+  "badges":{stylus:["bootstrap/bootstrap-stylus/bootstrap/badges.styl"]},
+  "jumbotron":{stylus:["bootstrap/bootstrap-stylus/bootstrap/jumbotron.styl"]},
+  "thumbnails":{stylus:["bootstrap/bootstrap-stylus/bootstrap/thumbnails.styl"]},
+  "alerts":{stylus:["bootstrap/bootstrap-stylus/bootstrap/alerts.styl"]},
+  "progress-bars":{stylus:["bootstrap/bootstrap-stylus/bootstrap/progress-bars.styl"]},
+  "media":{stylus:["bootstrap/bootstrap-stylus/bootstrap/media.styl"]},
+  "list-group":{stylus:["bootstrap/bootstrap-stylus/bootstrap/list-group.styl"]},
+  "panels":{stylus:["bootstrap/bootstrap-stylus/bootstrap/panels.styl"]},
+  "responsive-embed":{stylus:["bootstrap/bootstrap-stylus/bootstrap/responsive-embed.styl"]},
+  "wells":{stylus:["bootstrap/bootstrap-stylus/bootstrap/wells.styl"]},
+  "close":{stylus:["bootstrap/bootstrap-stylus/bootstrap/close.styl"]},
+
+  "component-animations":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/component-animations.styl"],
+    js:["bootstrap/bootstrap-stylus/js/transition.js"]
+  },
+  "dropdowns":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/dropdowns.styl"],
+    js:["bootstrap/bootstrap-stylus/js/dropdown.js"]
+  },
+  "modals":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/modals.styl"],
+    js:["bootstrap/bootstrap-stylus/js/transition.js", "bootstrap/bootstrap-stylus/js/modal.js"]
+  },
+  "tooltip":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/tooltip.styl"],
+    js:["bootstrap/bootstrap-stylus/js/tooltip.js"]
+  },
+  "popovers":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/tooltip.styl", "bootstrap/bootstrap-stylus/bootstrap/popovers.styl"],
+    js:["bootstrap/bootstrap-stylus/js/transition.js","bootstrap/bootstrap-stylus/js/tooltip.js","bootstrap/bootstrap-stylus/js/popover.js"]
+  },
+  "carousel":{
+    stylus:["bootstrap/bootstrap-stylus/bootstrap/carousel.styl"],
+    js:["bootstrap/bootstrap-stylus/js/transition.js", "bootstrap/bootstrap-stylus/js/carousel.js"]
+  },
+
+  "affix":{js:["bootstrap/bootstrap-stylus/js/affix.js"]},
+  "alert":{js:["bootstrap/bootstrap-stylus/js/alert.js"]},
+  "button":{js:["bootstrap/bootstrap-stylus/js/button.js"]},
+  "collapse":{js:["bootstrap/bootstrap-stylus/js/transition.js","bootstrap/bootstrap-stylus/js/collapse.js"]},
+  "scrollspy":{js:["bootstrap/bootstrap-stylus/js/scrollspy.js"]},
+  "tab":{js:["bootstrap/bootstrap-stylus/js/tab.js"]},
+  "transition":{js:["bootstrap/bootstrap-stylus/js/transition.js"]}
+};